test: test against redis in ci (#5)
Abstracts away the client initialization in the tests to initialize Momento vs Redis based on the environment variable TEST_REDIS. We add extra scripts to package.json to test against Momento vs Redis, add a Redis container to the build workflow, and add a job to test against Redis. We provide a sample devcontainer to code and debug against the Redis environment out of the box. Lastly we update the CONTRIBUTING file with the latest setup and test instructions.

---------

Co-authored-by: Matt Straathof <11823378+bruuuuuuuce@users.noreply.github.com>