chore: update README and errors to improve info about Redis API support
At present this library only supports a subset of the Redis APIs.
This PR updates the README to provide more info about which APIs are
supported and lets TypeScript users know that they can opt-in to our
interface in order to get compile-time checking as to whether we
support the APIs that they need.

We also add some contact info to the error message so that if users
call an API that we don't support yet, they will know how to reach
out to us to ask for us to support it.